PCB Layout
Created initial PCB layout and included more complicated connections to give an idea of what's going on. There's a connection that goes under the board since my brain does not know how to rearrange the components to work all on the front. Also, I made a board outline :D. Also this made me a bit crazy