Made Recommended changes
From savo:
C11 is a decoupling capacitor for one of the HX711s, it should be closer to it
You can get rid of two of the mounting holes. Two are enough.
Try to bring the USB stuff closer to the STM32, since it operates at high speed and we should keep the connections short.
Similarly try to bring the programming header closer to the STM32.
You can move the button stuff further from the STM32, those will be operated at kHz which is "slow"
The resistors you picked for the LEDs are a fair bit smaller than needed, 100+ Ohms is good.
No need for the pulldown R34
Power LEDs are cool, but maybe just one on the 3V3 rail is enough since thats the critical line. You can probably use the other one as a user configurable one too